services/worker/pyproject.toml
[project]
name = "lecpa-worker"
version = "0.1.0"
description = "Celery worker for document ingestion, OCR, and embeddings"
requires-python = ">=3.11"
dependencies = [
    "celery[redis]>=5.3.0",
    "redis>=5.0.0",
    "sqlalchemy>=2.0.25",
    "psycopg2-binary>=2.9.9",
    "pgvector>=0.2.4",
    "boto3>=1.34.0",
    "sentence-transformers>=2.3.0",
    "pymupdf>=1.23.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.2.0",
    "pdf2image>=1.17.0",
    "tenacity>=8.2.0",
    "structlog>=24.1.0",
    "tqdm>=4.66.0",
    "orjson>=3.9.0",
    "shared",
]

[tool.uv.sources]
shared = { workspace = true }

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tasks"]
